+++ templates/nodejs/src/index.js
@@ -1,0 +1,7 @@
+/**
+ * Entry point for the application
+ */
+
+// TODO: Implement main functionality
+
+module.exports = {};